requirements.txt
flask
flask-cors
flask-mail
flask-sqlalchemy
flask-limiter
pymysql
python-dotenv
email-validator 
dnspython
cryptography
redis